Convert music and sfx to Ogg Vorbis
This reduces the amount of I/O necessary. This avoids stutters when
streaming music on devices with slow I/O. Thanks to Paul Cercueil
for coming up with and testing this idea.

I put the conversion in the Makefile, so we have the option to switch
to a different codec later, or use an improved encoder or change the
quality setting.

The encoder is instructed to resample the audio if necessary. Some of
the sfx weren't at 44100 kHz and I imagine oggenc can do a better job
at resampling than SDL_mixer, since it doesn't run realtime.

I picked -q2 as the quality setting. In general -q4 to -q6 is
recommended for transparency, but with this particular music
-q2 sounded fine to me. Besides, reducing the chance of stutters
probably contributes more to a good audio experience than more
accurate encoding.

Note that despite the name, Mix_LoadWAV() loads Ogg files just fine.